make common patch for ellise and rectangle
